requirements.txt
# Web Framework & UI
streamlit==1.55.0
fastapi==0.115.0
python-dotenv==1.0.1
uvicorn==0.34.0

# Machine Learning & Data
pandas==2.2.3
numpy==2.1.0
scikit-learn==1.6.1
xgboost==2.1.3
joblib==1.4.2

# Generative AI & RAG (2026 Handshake)
langchain>=1.2.0
langchain-community>=0.4.0
langchain-google-genai>=2.1.0
chromadb==1.5.5
pypdf==5.2.0
sentence-transformers==3.2.0

# Databases
psycopg2-binary==2.9.10
sqlalchemy==2.0.48

# Scraper
seleniumbase==4.47.1